agentcore-demo/requirements.txt
bedrock-agentcore>=1.0.3
strands-agents>=1.13.0
aws-opentelemetry-distro>=0.10.0
python-dotenv>=1.2.1
